separated ui message types
